Add: downloader check if the file already exists

+++ SoundcloudDownloader/downloader.py
@@ -136,6 +136,9 @@ class SCDL:
         else:
             path = path / filename
 
+        if os.path.exists(path):
+            raise FileExistsError('File already exists')
+
         if metadata['protocol'] == "direct" or metadata['protocol'] == "stream":
             with open(path, 'wb') as f:
                 for chunk in self.session.get(metadata['download_url'], stream=True):
